do not create soft links to non-existent binaries

+++ Makefile
@@ -22,8 +22,8 @@ install:
 	ln -sf $(INSTALLDIR)/bin/etraining /usr/local/bin/etraining
 	ln -sf $(INSTALLDIR)/bin/prepareAlign /usr/local/bin/prepareAlign
 	ln -sf $(INSTALLDIR)/bin/fastBlockSearch /usr/local/bin/fastBlockSearch
-	ln -sf $(INSTALLDIR)/bin/load2db /usr/local/bin/load2db
-	ln -sf $(INSTALLDIR)/bin/getSeq /usr/local/bin/getSeq
+	if [ -f $(INSTALLDIR)/bin/load2db ] ; then ln -sf $(INSTALLDIR)/bin/load2db /usr/local/bin/load2db ; fi	
+	if [ -f $(INSTALLDIR)/bin/getSeq ] ; then ln -sf $(INSTALLDIR)/bin/getSeq /usr/local/bin/getSeq ; fi
 	
 # for internal purposes:
 release:
